Use reflection only assembly context when cleaning up add-ins
Fixes #6
